refactor(avatars): drop orphaned app-avatar: symbolic-ref machinery
The app-avatar:<id> symbolic pointer was meant to power curated,
selectable avatar sets — users pick an agent avatar by symbolic ID
instead of baking in one-off URLs. But the resolver half was never
built / was removed in #1132 (the goose-set collections). What
survived was only the parsing + plumbing half: parseGooseAppAvatarId
was only ever called by its own validators, and no id->image map
existed anywhere. The machinery stored and round-tripped an ID that
nothing could render.

This removes that orphaned plumbing:
- delete the gooseAppAvatarRefs module (parser + validators + tests)
- collapse the profile/hooks.ts skip-fetch guard to a plain snapshot
- inline personaDialogState.ts import resolution to keep persistable
  http(s):/data:/blob: URLs only (byte-for-byte identical branch;
  symbolic refs now drop instead of round-trip — intended, nothing
  renders them)
- retire the dead app-avatar: vocabulary in Rust test fixtures

The Rust avatar_ref field stays: it is a generic opaque carrier of the
avatar: frontmatter value (it holds real URLs), not specific to
app-avatar:.

See #1132 for the breadcrumb to re-add selectable avatar sets — that is
where the resolver + picker UI should be rebuilt alongside this plumbing.
